🚚 약관 내용은 entities에서 상수로 관리
