--- src/components/modal/ServiceTerm.tsx
@@ -1,96 +1,0 @@
-import { Button } from '@/components/ui/button';
-import { ModalSelectBackground } from '@/components/ui/layout';
-import { MarkdownPreview } from '@/components/ui/markdown-preview';
-import { useDialog } from '@/shared/modal/hooks';
-
-export const ServiceTermModal = ({ onClose }: { onClose: () => void }) => {
-  const { isVisible, handleClose } = useDialog({
-    onClose,
-  });
-
-  return (
-    <ModalSelectBackground
-      variant="transparent"
-      isVisible={isVisible}
-      onOutSlideClick={handleClose}
-      className="max-h-screen w-full max-w-[800px] overflow-y-scroll rounded-none text-left"
-    >
-      <div className="flex flex-col pt-4">
-        <MarkdownPreview content={term} />
-      </div>
-      <div className="flex flex-col-reverse gap-2 sm:flex-row">
-        <Button variant="secondary" onClick={handleClose} className="flex-1">
-          닫기
-        </Button>
-      </div>
-    </ModalSelectBackground>
-  );
-};
-
-const term = `
-# 인턴하샤 서비스 이용약관
-
-## 제1조 (목적)  
-이 약관은 인턴하샤(이하 ‘회사’라 합니다)가 제공하는 무료 서비스의 이용과 관련하여 회사와 회원 간의 권리, 의무 및 책임사항, 기타 필요한 사항을 규정함을 목적으로 합니다.
-
-## 제2조 (약관의 게시와 개정)  
-1. 회사는 이 약관을 회원이 그 전부를 확인할 수 있도록 기술적 조치를 취합니다.  
-2. 회사는 회원이 약관에 동의하기에 앞서 청약철회 제한 등 중요한 내용을 회원이 쉽게 이해할 수 있도록 별도의 화면에 제공합니다.  
-3. 회사는 관련 법령을 위배하지 않는 범위에서 이 약관을 개정할 수 있습니다.  
-4. 개정 시에는 적용일자 및 개정사유를 명시하여 7일 전부터 홈페이지에 공지합니다.  
-5. 회원이 개정약관에 대해 명시적으로 거부하지 않는 한, 동의한 것으로 봅니다.  
-6. 회원이 개정약관에 동의하지 않을 경우 이용계약을 해지할 수 있습니다.  
-7. 회원은 해지로 인해 손해가 발생한 경우 회사에 손해배상을 청구할 수 있습니다.  
-
-## 제3조 (약관의 해석)  
-이 약관에서 정하지 아니한 사항에 대해서는 관계 법령, 관례 및 서비스 운영정책 등에 따릅니다.
-
-## 제4조 (회원에 대한 통지)  
-1. 회사는 회원에게 서비스 화면, 전자우편 등으로 개별 통지할 수 있습니다.  
-2. 회원 전체에 대한 통지는 7일 이상 게시함으로써 개별 통지에 갈음할 수 있습니다. 단, 개별 통지가 필요한 사항은 별도로 통지합니다.
-
-## 제5조 (서비스 내용 등의 게시)  
-1. 회사는 아래 사항을 서비스 초기화면 또는 FAQ 등을 통해 회원이 알기 쉽게 제공합니다.  
-   - **서비스 명칭**: 인턴하샤  
-   - **제공자**: 인턴하샤  
-   - **제공 서비스**: 기업 인턴 공고 제공, 커피챗 연결, 프로필 작성  
-   - **이용 조건**: 무료 제공, 서울대학교 재학생 인증 필요  
-2. 서비스 제공을 위한 최소한의 기술사양은 별도 안내에 따릅니다.  
-3. 서비스는 전적으로 무료이며, 대금 환불 등의 절차는 해당되지 않습니다.
-
-## 제6조 (이용계약의 성립)  
-1. 회원은 다음 절차에 따라 이용을 신청합니다.  
-   - 서비스 회원가입 신청  
-   - 서울대학교 이메일(@snu.ac.kr) 인증  
-   - 약관 동의 및 기타 정보 입력  
-2. 회사는 아래의 경우 이용을 제한하거나 거부할 수 있습니다.  
-   - 실명이 아닌 경우 또는 타인의 명의를 사용한 경우  
-   - 허위의 정보를 기재한 경우  
-   - 서울대학교 학생이 아님에도 불구하고 인증을 시도한 경우  
-   - 기타 서비스 운영에 지장이 발생할 수 있는 경우  
-3. 이용계약의 성립은 가입 완료 시점으로 합니다.
-
-## 제7조 (서비스 제공)  
-회사가 제공하는 서비스는 다음과 같습니다.  
-- **기업 인턴 공고 제공**: 기업명, 위치, 도메인, 연봉, 직무 설명, 기업 소개 등 정보 제공  
-- **기업 대상 커피챗 연결**: 회원이 신청한 커피챗을 기업이 수락 시 이메일 전달  
-- **프로필 작성**: 학과, 학번, 이력서, 희망 직군 등 정보를 기반으로 기업에 프로필 제공  
-
-> 작성된 프로필 및 게시물은 커피챗 연결 목적 외에 다른 용도로 사용되지 않습니다.
-
-## 제8조 (서비스 이용 제한)  
-회원이 서울대학교 학생이 아님에도 불구하고 허위로 인증한 경우, 회사는 해당 회원에 대해 서비스 이용을 제한하거나 계약을 해지할 수 있습니다.
-
-## 제9조 (책임 제한)  
-1. 회사는 천재지변, 기술적 장애 등 불가항력으로 인한 서비스 중단에 대해 책임을 지지 않습니다.  
-2. 회원의 귀책사유로 인한 서비스 이용 장애에 대해서는 회사가 책임지지 않습니다.  
-3. 회사는 회원 간, 또는 회원과 제3자 간에 발생한 분쟁에 대하여 개입하거나 책임을 지지 않습니다.
-
-## 제10조 (분쟁의 해결)  
-서비스 이용과 관련하여 분쟁이 발생할 경우, 회사의 소재지인 와플스튜디오 관할 법원을 제1심 관할 법원으로 합니다.
-
-<br/>
-
-**[부칙]**  
-이 약관은 2025년 5월 8일부터 시행합니다.
-`;
