+++ src/entities/agreements.ts
@@ -22,3 +22,150 @@ export const TERMS = [
     url: PATH.INFORMATION_TERM,
   },
 ];
+
+export const SERVICE_TERM_CONTENT = `
+# 인턴하샤 서비스 이용약관
+
+## 제1조 (목적)  
+이 약관은 인턴하샤(이하 ‘회사’라 합니다)가 제공하는 무료 서비스의 이용과 관련하여 회사와 회원 간의 권리, 의무 및 책임사항, 기타 필요한 사항을 규정함을 목적으로 합니다.
+
+## 제2조 (약관의 게시와 개정)  
+1. 회사는 이 약관을 회원이 그 전부를 확인할 수 있도록 기술적 조치를 취합니다.  
+2. 회사는 회원이 약관에 동의하기에 앞서 청약철회 제한 등 중요한 내용을 회원이 쉽게 이해할 수 있도록 별도의 화면에 제공합니다.  
+3. 회사는 관련 법령을 위배하지 않는 범위에서 이 약관을 개정할 수 있습니다.  
+4. 개정 시에는 적용일자 및 개정사유를 명시하여 7일 전부터 홈페이지에 공지합니다.  
+5. 회원이 개정약관에 대해 명시적으로 거부하지 않는 한, 동의한 것으로 봅니다.  
+6. 회원이 개정약관에 동의하지 않을 경우 이용계약을 해지할 수 있습니다.  
+7. 회원은 해지로 인해 손해가 발생한 경우 회사에 손해배상을 청구할 수 있습니다.  
+
+## 제3조 (약관의 해석)  
+이 약관에서 정하지 아니한 사항에 대해서는 관계 법령, 관례 및 서비스 운영정책 등에 따릅니다.
+
+## 제4조 (회원에 대한 통지)  
+1. 회사는 회원에게 서비스 화면, 전자우편 등으로 개별 통지할 수 있습니다.  
+2. 회원 전체에 대한 통지는 7일 이상 게시함으로써 개별 통지에 갈음할 수 있습니다. 단, 개별 통지가 필요한 사항은 별도로 통지합니다.
+
+## 제5조 (서비스 내용 등의 게시)  
+1. 회사는 아래 사항을 서비스 초기화면 또는 FAQ 등을 통해 회원이 알기 쉽게 제공합니다.  
+   - **서비스 명칭**: 인턴하샤  
+   - **제공자**: 인턴하샤  
+   - **제공 서비스**: 기업 인턴 공고 제공, 커피챗 연결, 프로필 작성  
+   - **이용 조건**: 무료 제공, 서울대학교 재학생 인증 필요  
+2. 서비스 제공을 위한 최소한의 기술사양은 별도 안내에 따릅니다.  
+3. 서비스는 전적으로 무료이며, 대금 환불 등의 절차는 해당되지 않습니다.
+
+## 제6조 (이용계약의 성립)  
+1. 회원은 다음 절차에 따라 이용을 신청합니다.  
+   - 서비스 회원가입 신청  
+   - 서울대학교 이메일(@snu.ac.kr) 인증  
+   - 약관 동의 및 기타 정보 입력  
+2. 회사는 아래의 경우 이용을 제한하거나 거부할 수 있습니다.  
+   - 실명이 아닌 경우 또는 타인의 명의를 사용한 경우  
+   - 허위의 정보를 기재한 경우  
+   - 서울대학교 학생이 아님에도 불구하고 인증을 시도한 경우  
+   - 기타 서비스 운영에 지장이 발생할 수 있는 경우  
+3. 이용계약의 성립은 가입 완료 시점으로 합니다.
+
+## 제7조 (서비스 제공)  
+회사가 제공하는 서비스는 다음과 같습니다.  
+- **기업 인턴 공고 제공**: 기업명, 위치, 도메인, 연봉, 직무 설명, 기업 소개 등 정보 제공  
+- **기업 대상 커피챗 연결**: 회원이 신청한 커피챗을 기업이 수락 시 이메일 전달  
+- **프로필 작성**: 학과, 학번, 이력서, 희망 직군 등 정보를 기반으로 기업에 프로필 제공  
+
+> 작성된 프로필 및 게시물은 커피챗 연결 목적 외에 다른 용도로 사용되지 않습니다.
+
+## 제8조 (서비스 이용 제한)  
+회원이 서울대학교 학생이 아님에도 불구하고 허위로 인증한 경우, 회사는 해당 회원에 대해 서비스 이용을 제한하거나 계약을 해지할 수 있습니다.
+
+## 제9조 (책임 제한)  
+1. 회사는 천재지변, 기술적 장애 등 불가항력으로 인한 서비스 중단에 대해 책임을 지지 않습니다.  
+2. 회원의 귀책사유로 인한 서비스 이용 장애에 대해서는 회사가 책임지지 않습니다.  
+3. 회사는 회원 간, 또는 회원과 제3자 간에 발생한 분쟁에 대하여 개입하거나 책임을 지지 않습니다.
+
+## 제10조 (분쟁의 해결)  
+서비스 이용과 관련하여 분쟁이 발생할 경우, 회사의 소재지인 와플스튜디오 관할 법원을 제1심 관할 법원으로 합니다.
+
+<br/>
+
+**[부칙]**  
+이 약관은 2025년 5월 8일부터 시행합니다.
+`;
+
+export const INFO_TERM_CONTENT = `
+# 개인정보 처리방침
+
+인턴하샤(이하 '회사')는 『개인정보 보호법』 등 관련 법령상의 개인정보 보호 규정을 준수하며, 이용자의 개인정보 보호를 위해 최선을 다하고 있습니다. 본 개인정보 처리방침은 회사가 제공하는 서비스와 관련하여 이용자의 개인정보를 어떻게 수집, 이용, 보관, 파기하는지를 설명하며, **2025년 5월 8일부터 적용**됩니다.
+
+## 1. 수집하는 개인정보
+
+회사는 다음과 같은 개인정보를 수집합니다.
+
+- **학생 회원가입**
+  - 수집 항목: 실명, 서울대학교 이메일 주소(@snu.ac.kr)
+  - 수집 방식: 회원가입 시 직접 입력 또는 학교 메일 인증
+
+- **회사 회원가입**
+  - 수집 항목: 회사명, 회사 이메일 주소
+  - 수집 방식: 이메일을 통한 개별 수집
+
+- **서비스 이용**
+  - 수집 항목: 학번, 학과, 이력서, 포트폴리오
+  - 수집 방식: 서비스 이용 중 이용자가 직접 입력
+
+## 2. 개인정보의 이용 목적
+
+회사는 수집한 개인정보를 다음과 같은 목적을 위해 활용합니다.
+
+- 회원 식별 및 인증, 서비스 제공 및 유지·관리  
+- 인턴 커피챗 신청 및 매칭  
+- 신규 서비스 개발 및 기존 서비스 개선  
+- 학생이 신청한 커피챗이 수락된 경우, **실명, 이메일 등의 정보는 신청한 회사에 전달됨**
+
+> ※ 단, 회사에 학생 정보 제공을 원하지 않는 경우 커피챗 신청을 취소할 수 있습니다.
+
+## 3. 개인정보의 보유 및 이용기간
+
+- **학생 회원**: 탈퇴 시점에 모든 개인정보가 즉시 파기됩니다.  
+- **회사 회원**: 탈퇴 요청 시 회사가 수동으로 탈퇴를 처리하며, **회사 공고 이외의 모든 데이터는 삭제**됩니다.
+
+## 4. 개인정보의 제3자 제공
+
+회사는 원칙적으로 이용자의 개인정보를 제3자에게 제공하지 않습니다.  
+단, 아래의 경우 예외로 합니다.
+
+- **이용자가 커피챗을 신청하고, 해당 신청이 수락된 경우**:  
+  - 학생의 실명, 이메일 등 일부 정보가 신청한 회사에 제공됩니다.
+
+## 5. 개인정보 처리의 위탁
+
+회사는 개인정보 처리를 외부에 위탁하지 않습니다.
+
+## 6. 개인정보의 파기
+
+회사는 개인정보 보유 기간의 경과 또는 처리 목적 달성 시 지체 없이 해당 개인정보를 파기합니다.
+
+- **전자적 파일**: 복구 및 재생이 불가능한 방식으로 삭제  
+- **종이 문서**: 분쇄 또는 소각
+
+## 7. 만 14세 미만 아동의 개인정보
+
+회사는 만 14세 미만 아동으로부터 개인정보를 수집하지 않으며, 해당 연령층의 가입을 불허합니다.
+
+## 8. 이용자의 권리와 행사 방법
+
+- 이용자는 언제든지 본인의 개인정보에 대해 **열람, 정정, 삭제, 처리정지**를 요청할 수 있습니다.  
+- 동의를 철회하고자 하는 경우, 서비스 내 **탈퇴 기능**을 통해 개인정보의 수집·이용을 철회할 수 있습니다.
+
+## 9. 개인정보 보호책임자 안내
+
+회사는 개인정보 보호에 관한 업무를 총괄해서 책임지고 있으며, 개인정보 관련 문의사항에 신속하고 성실하게 답변해드리고 있습니다.
+
+- **개인정보 보호책임자**: 김연우  
+- **이메일**: internhasha.official@gmail.com
+
+## 10. 고지의 의무
+
+이 개인정보 처리방침은 **2025년 5월 8일부터 적용**됩니다.  
+변경이 있을 경우 사전 공지를 통해 안내드립니다.
+
+`;
