+++ src/feature/auth/ui/signUp/LocalSignUpForm.tsx
@@ -3,8 +3,8 @@ import { useState } from 'react';
 import { TermsAgreementField } from '@/components/field/TermsAgreeField';
 import { FormContainer } from '@/components/form/FormContainer';
 import { LabelContainer } from '@/components/label/LabelContainer';
-import { InfoTermModal } from '@/components/modal/InformationTerm';
-import { ServiceTermModal } from '@/components/modal/ServiceTerm';
+import { InfoTermModal } from '@/components/modal/InformationTermModal';
+import { ServiceTermModal } from '@/components/modal/ServiceTermModal';
 import { ProgressBar } from '@/components/progressBar/ProgressBar';
 import { FormErrorResponse } from '@/components/response/formResponse';
 import { Button } from '@/components/ui/button';
